add str2ts tool

+++ Makefile
@@ -9,19 +9,20 @@ rmdeps:
 
 self:   prep
 	if test ! -d src; then mkdir src; fi
-	if -d vendor/src; then cp -r vendor/src/* src/; fi
+	if test -d vendor; then cp -r vendor/* src/; fi
 
 deps:   
 
 vendor-deps: rmdeps deps
 	if test ! -d vendor; then mkdir vendor; fi
-	if test -d vendor/src; then rm -rf vendor/src; fi
-	cp -r src vendor/src
-	find vendor -name '.git' -print -type d -exec rm -rf {} +
-	rm -rf src
+	if test -d vendor; then rm -rf vendor; fi
+	if test -d src; then cp -r src vendor; fi
+	if test -d vendor; then find vendor -name '.git' -print -type d -exec rm -rf {} +; fi
+	if test -d src; then rm -rf src; fi
 
 fmt:
 	go fmt cmd/*.go
 
 bin:	self
 	@GOPATH=$(GOPATH) go build -o bin/ts2str cmd/ts2str.go
+	@GOPATH=$(GOPATH) go build -o bin/str2ts cmd/str2ts.go
